+++ src/main/java/blackjack/domain/GameResults.java
@@ -1,5 +1,7 @@
 package blackjack.domain;
 
+import blackjack.domain.player.Player;
+
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
