+++ src/main/java/blackjack/domain/player/Dealer.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 public class Dealer extends Player {
     public Dealer() {
