+++ src/main/java/blackjack/domain/player/Gambler.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 public class Gambler extends Player {
     public Gambler(String name) {
