+++ src/test/java/blackjack/domain/GameManagerTest.java
@@ -1,5 +1,7 @@
 package blackjack.domain;
 
+import blackjack.domain.card.BlackjackShuffle;
+import blackjack.domain.player.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
